test(LadderGameTest): Ladder 생성 테스트 코드를 LadderGameTest 클래스로 이관
